requirements.txt
certifi==0.0.8
chardet==1.0.1
coverage==4.1
coveralls==1.1
disqus==0.0.4
dj-database-url==0.4.1
Django==1.9.7
django-appconf==1.0.2
django-crispy-forms==1.6.0
django-forms-bootstrap==3.0.1
django-formtools==1.0
django-jsonfield==1.0.0
django-messages==0.5.3
django-model-utils==2.5
django-nose==1.4.4
django-notifications-hq==1.2
django-notify-x==0.1.6
django-pagedown==0.1.1
django-registration-redux==1.4
docopt==0.6.2
Flask==0.8
funcsigs==1.0.2
gunicorn==0.14.2
Jinja2==2.6
jsonfield==1.0.3
markerlib==0.6.0
mock==2.0.0
nose==1.3.7
pbr==1.10.0
pbs==0.110
Pillow==3.2.0
psycopg2==2.6.2
pusher==1.1.3
pytz==2016.4
requests==2.10.0
six==1.10.0
Werkzeug==0.8.3
